back/requirements.txt
Flask 
Flask-Cors
pymongo
python-dotenv
PyJWT
